scraper/requirements.txt
psycopg2-binary
geopy
python-dotenv
pydantic-settings
tenacity  